stock-sentiment-app/github stock porject/StockSentimentscreener-Krishna/stock-sentiment-app/requirements.txt
Flask
requests
yfinance
textblob
plotly
python-dotenv
pandas
